requirements.txt
aiohttp>=3.9
cryptography>=42.0
python-dotenv>=1.0
discord.py>=2.3
supabase>=2.0
fastapi>=0.110
uvicorn>=0.27
slowapi>=0.1.9
